requirements.txt
datasets~=3.0.2
h5py~=3.12.1
llama-cpp-python~=0.3.1
numpy~=2.1.2
optuna~=4.0.0
psutil~=6.1.0
pydantic~=2.9.2
PyYAML~=6.0.2
scipy~=1.14.1
tdigest~=0.5.2.2
torch~=2.5.1
